Refactor for using MCD from test_bench
