+++ mcd/TestMonteCarloDropoutModel.py
@@ -1,5 +1,8 @@
 import unittest
 
+from test_bench.model import MonteCarloDropoutModel
+
+
 class TestMonteCarloDropoutModel(unittest.TestCase):
 
     def setUp(self):
@@ -11,7 +14,8 @@ class TestMonteCarloDropoutModel(unittest.TestCase):
         self.assertIsNotNone(self.model.get_generator())
 
     def test_checkpoint_name(self):
-        self.assertEqual(self.model._get_checkpoint_name(), 'checkpoint_muon')
+        self.assertEqual(self.model._get_checkpoint_name(),
+                         'checkpoint_dir')
 
     def test_model_string_representation(self):
         self.assertEqual(str(self.model), 'muon_0.1')
@@ -21,6 +25,7 @@ class TestMonteCarloDropoutModel(unittest.TestCase):
             self.model._restore_model()
         except Exception as e:
             self.fail(f"Model restore failed: {e}")
+
 
 if __name__ == '__main__':
     unittest.main()
